[ProposePage] move proposal page to its own subdirectory
